refactor: separate data loading and validation
